Utilities/requirements.txt
Pillow==10.4.0
tqdm==4.67.0 